requirements.txt
legobot>=1.0.1,<=2.0.0
python-whois
